backend/requirements.txt
fastapi
pydantic
uvicorn
requests
pydantic
shapely
geopy
pyyaml
pvlib
pandas
joblib
scikit-learn